removed non-callback functions from main.py
