+++ pemfc_dash/layout_functions.py
@@ -3,6 +3,8 @@ from dash import html
 from dash import dcc
 import copy
 
+from . import data_conversion as dc
+
 ID_LIST = []  # Keep track with generated IDs
 CONTAINER_LIST = []
 
@@ -461,6 +463,19 @@ def row_input(label='', ids='', value='', type='', dimensions='', options='',
         inputs = html.Div()
 
     return inputs
+
+
+def conditional_dropdown_menu(dropdown_value, data):
+    results = dc.read_data(data)
+    result_set = results.iloc[0]
+    local_data = result_set["local_data"]
+    if 'value' in local_data[dropdown_value]:
+        return [], None, {'visibility': 'hidden'}
+    else:
+        options = [{'label': key, 'value': key} for key in
+                   local_data[dropdown_value]]
+        value = options[0]['value']
+        return options, value, {'visibility': 'visible'}
 
 
 graph_font_props = {'small': {'size': 12, 'color': 'black', 'family': 'Arial'},
